requirements.txt
fastapi 
uvicorn[standard] 
motor 
Pillow 
numpy 
python-multipart 
gradio 
requests 
insightface
opencv-python-headless
onnxruntime
mysql-connector-python
python-dotenv
